pyproject.toml
[project]
name = "auto-sre"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.9.0",
    "asyncpg>=0.31.0",
    "bcrypt==4.0.1",
    "docker>=7.1.0",
    "fastapi>=0.124.0",
    "jieba>=0.42.1",
    "langchain>=1.1.2",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.1",
    "langchain-huggingface>=1.1.0",
    "langchain-openai>=1.1.0",
    "langgraph>=1.0.4",
    "minio>=7.2.20",
    "numpy>=2.3.5",
    "paramiko>=4.0.0",
    "passlib>=1.7.4",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pypdf2>=3.0.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=5.0.0",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.45",
    "streamlit>=1.52.1",
    "uvicorn>=0.38.0",
]
